Moved type from service into types, fixed format
