+++ libs/api/domains/content-search/src/lib/contentSearch.resolver.ts
@@ -7,7 +7,7 @@ import { WebSearchAutocomplete } from './models/webSearchAutocomplete.model'
 
 @Resolver()
 export class ContentSearchResolver {
-  constructor(private contentSearchService: ContentSearchService) { }
+  constructor(private contentSearchService: ContentSearchService) {}
 
   @Query(() => SearchResult)
   searchResults(@Args('query') query: SearcherInput): Promise<SearchResult> {
