+++ libs/content-search-toolkit/src/types/search.ts
@@ -1,4 +1,4 @@
-import { elasticTagField } from '.';
+import { elasticTagField } from '.'
 
 export interface SearchInput {
   queryString: string
