+++ libs/content-search-toolkit/src/types/tagAggregation.ts
@@ -3,3 +3,27 @@ export interface TagAggregationInput {
   tagType: string
   size?: number
 }
+
+export interface TagAggregationResponse {
+  group: {
+    filtered: {
+      doc_count: number
+      count: {
+        buckets: [
+          {
+            key: string
+            doc_count: number
+            value: {
+              buckets: [
+                {
+                  key: string
+                  doc_count: number
+                },
+              ]
+            }
+          },
+        ]
+      }
+    }
+  }
+}
